# swagger.yaml
openapi: 3.0.3
info:
  title: Dating App API
  version: 1.0.0
  description: |
    Music-based dating application backend. Matches users based on Spotify listening history,
    lifestyle compatibility, and behavioral learning.

    ## Authentication
    Most endpoints require a JWT passed as `Authorization: Bearer <token>`.
    JWT claims: `{ userId, tokenVersion, sub, iat, exp }`.

    After a password reset `tokenVersion` is incremented — all prior tokens are invalidated.
    Deleted-user JWTs return **401** (not 404).

    ## Email Verification
    Unverified-email users receive **403** `{"error":"Email verification required"}` on all
    endpoints except `/api/v1/auth/**`, `/api/v1/onboarding/**`, `/public/**`, and `/api/test/**`.

    ## Rate Limiting
    | Action | Limit | Key |
    |--------|-------|-----|
    | Swipe | 60/min | userId |
    | Score calculation | 30/min | userId |
    | Unauthenticated | per IP | remoteAddr |

    Exceeded limits return **429**.

    ## Error Contract
    Every error response follows:
    ```json
    { "code": "string", "message": "string", "fields": {}, "timestamp": "ISO8601" }
    ```
    Always switch on `code`, never on `message`.

  contact:
    name: API Support
    email: mladen.hangi@kodelab.hr
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register with email and password
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: alice@example.com
              name: Alice Smith
              password: Secret@123
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9...
                userId: 550e8400-e29b-41d4-a716-446655440000
                email: alice@example.com
                name: Alice Smith
                registrationStage: STARTED
                emailVerified: false
                authProvider: EMAIL
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/EmailExists'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: alice@example.com
              password: Secret@123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/AccountLocked'
  /api/v1/auth/spotify-login:
    post:
      tags: [Authentication]
      summary: Spotify OAuth login — creates or finds user
      operationId: spotifyLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotifyLoginRequest'
      responses:
        '200':
          description: Spotify login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email address with token
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Email verified successfully
        '400':
          $ref: '#/components/responses/InvalidToken'
  /api/v1/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Resend email verification link
      description: Always returns 200 to prevent user enumeration.
      operationId: resendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Verification email sent (or silently skipped if email not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: If the email exists, verification email will be sent
  /api/v1/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset link
      description: Always returns 200 to prevent user enumeration.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Reset link sent (or silently skipped)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: If the email exists, a password reset link will be sent
  /api/v1/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password using a reset token
      description: |
        Resets the password and increments `tokenVersion`, invalidating all existing JWTs.
        The frontend must force re-login after a successful reset.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Password reset successfully
        '400':
          $ref: '#/components/responses/InvalidToken'
  /api/v1/auth/connect-spotify:
    post:
      tags: [Authentication]
      summary: Link Spotify to an existing email account
      operationId: connectSpotify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectSpotifyRequest'
      responses:
        '200':
          description: Spotify account linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  # ─── Onboarding ────────────────────────────────────────────────────────────
...
  /api/v1/onboarding/privacy-settings:
    put:
      tags: [Onboarding]
      summary: Update privacy settings
      operationId: updatePrivacySettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacySettingsRequest'
      responses:
        '200':
          description: Privacy settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDtoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Onboarding]
      summary: Get privacy settings (deprecated)
      description: '**Deprecated.** Use `GET /api/v1/onboarding/profile` instead. Sunset: 2026-06-11.'
      operationId: getPrivacySettings
      deprecated: true
      responses:
        '200':
          description: Privacy settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacySettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/users/me:
    delete:
      tags: [Account]
      summary: Delete the current user's account
      description: |
        Permanently deletes the account and all associated data.
        - EMAIL auth users must supply their password.
        - Spotify-only users may omit the request body.
        - All existing JWTs for this user will return 401 after deletion.
      operationId: deleteAccount
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
            example:
              password: Secret@123
      responses:
        '204':
          description: Account deleted — no content
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/{id}/profile:
    get:
      tags: [Account]
      summary: Get a user's public profile
      description: |
        Returns the public profile of another user.
        - If `isProfilePublic=false` → 404 (prevents user-ID enumeration).
        - If a block relationship exists (either direction) → 404.
        - Profile owner always sees full data; non-owners see privacy-gated fields.
      operationId: getUserProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: User public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── User / Spotify (authenticated) ───────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT with claims: `{ userId, tokenVersion, sub, iat, exp }`.
        `tokenVersion` is required — tokens without it are rejected with 401.

  # ─── Responses ─────────────────────────────────────────────────────────────
  responses:
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              email: must not be blank
              password: Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character
            timestamp: '2026-05-12T10:15:30.123Z'
...
    InvalidToken:
      description: Expired or malformed verification/reset token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_TOKEN
            message: Token is expired or invalid
            timestamp: '2026-05-12T10:15:30.123Z'
    Unauthorized:
      description: Missing or invalid JWT, or wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid email or password
            timestamp: '2026-05-12T10:15:30.123Z'
...
    EmailExists:
      description: Email address already registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: EMAIL_EXISTS
            message: An account with this email already exists
            timestamp: '2026-05-12T10:15:30.123Z'
    Conflict:
      description: Resource conflict — Spotify already connected or generic data conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: A conflict occurred. The resource may already exist.
            timestamp: '2026-05-12T10:15:30.123Z'
...
    AccountLocked:
      description: Too many failed login attempts
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACCOUNT_LOCKED
            message: Account is temporarily locked due to too many failed login attempts
            timestamp: '2026-05-12T10:15:30.123Z'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: User not found
            timestamp: '2026-05-12T10:15:30.123Z'
...
  schemas:
...
    Gender:
      type: string
      enum: [MALE, FEMALE, NON_BINARY, OTHER, PREFER_NOT_TO_SAY]
    SexualOrientation:
      type: string
      enum: [STRAIGHT, GAY, LESBIAN, BISEXUAL, PANSEXUAL, ASEXUAL, DEMISEXUAL, QUEER, QUESTIONING, OTHER]
    RegistrationStage:
      type: string
      enum:
        - STARTED
        - BASIC_PROFILE
        - LOCATION_INFO
        - PHOTOS
        - MUSIC_PREFERENCES
        - LIFESTYLE
        - PERSONALITY
        - DATING_PREFERENCES
        - PRIVACY_SETTINGS
        - FINISHED
...
    KidsPreference:
      type: string
      enum: [WANTS_KIDS, DOESNT_WANT_KIDS, HAS_KIDS, HAS_KIDS_WANTS_MORE, HAS_KIDS_DOESNT_WANT_MORE, OPEN_TO_KIDS, NOT_SURE, PREFER_NOT_TO_SAY]
    SmokingHabits:
      type: string
      enum: [NON_SMOKER, SOCIAL_SMOKER, REGULAR_SMOKER, TRYING_TO_QUIT, PREFER_NOT_TO_SAY]
    DrinkingHabits:
      type: string
      enum: [NON_DRINKER, SOCIAL_DRINKER, MODERATE_DRINKER, REGULAR_DRINKER, PREFER_NOT_TO_SAY]
    ExerciseFrequency:
      type: string
      enum: [NEVER, RARELY, ONCE_A_WEEK, FEW_TIMES_A_WEEK, DAILY, MULTIPLE_TIMES_DAILY]
    EducationLevel:
      type: string
      enum: [HIGH_SCHOOL, SOME_COLLEGE, ASSOCIATES_DEGREE, BACHELORS_DEGREE, MASTERS_DEGREE, DOCTORATE, TRADE_SCHOOL, PREFER_NOT_TO_SAY]
    MBTI:
      type: string
      enum: [ISFJ, ISFP, ISTJ, ISTP, INFJ, INFP, INTJ, INTP, ESFJ, ESFP, ESTJ, ESTP, ENFJ, ENFP, ENTJ, ENTP]
...
    Religion:
      type: string
      enum: [CHRISTIAN, MUSLIM, JEWISH, HINDU, BUDDHIST, SIKH, ATHEIST, AGNOSTIC, SPIRITUAL, OTHER, PREFER_NOT_TO_SAY]
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code. Always switch on this, never on message.
          example: VALIDATION_ERROR
          enum:
            - VALIDATION_ERROR
            - INVALID_ARGUMENT
            - INVALID_TOKEN
            - UNAUTHORIZED
            - FORBIDDEN
            - ACCOUNT_LOCKED
            - NOT_FOUND
            - EMAIL_EXISTS
            - CONFLICT
            - DUPLICATE_SWIPE
            - CONCURRENT_MODIFICATION
            - SPOTIFY_TOKEN_EXPIRED
            - SPOTIFY_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable description
        fields:
          type: object
          nullable: true
          description: Per-field validation messages. Only present for VALIDATION_ERROR.
          additionalProperties:
            type: string
          example:
            email: must not be blank
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp

    # ── Auth ───────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Alice Smith
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Must contain uppercase, lowercase, digit, and special character (@$!%*?&#)
          example: Secret@123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: Secret@123
    SpotifyLoginRequest:
      type: object
      required: [spotifyId, email, name, spotifyAccessToken, spotifyRefreshToken, spotifyTokenExpiresAt]
      properties:
        spotifyId:
          type: string
          example: spotify:user:alice123
        email:
          type: string
          format: email
          example: alice@example.com
        name:
          type: string
          example: Alice Smith
        spotifyAccessToken:
          type: string
          example: BQD...token
        spotifyRefreshToken:
          type: string
          example: AQD...refresh
        spotifyTokenExpiresAt:
          type: integer
          format: int64
          description: Epoch milliseconds
          example: 1747000000000
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: https://i.scdn.co/image/abc123
    EmailVerificationRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
          description: Same rules as registration password
          example: NewSecret@456
    ConnectSpotifyRequest:
      type: object
      required: [spotifyId, spotifyAccessToken, spotifyRefreshToken, spotifyTokenExpiresAt]
      properties:
        spotifyId:
          type: string
          maxLength: 255
          example: spotify:user:alice123
        spotifyAccessToken:
          type: string
          maxLength: 2048
        spotifyRefreshToken:
          type: string
          maxLength: 2048
        spotifyTokenExpiresAt:
          type: integer
          format: int64
          description: Epoch milliseconds
          example: 1747000000000
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT for subsequent requests
          example: eyJhbGciOiJIUzI1NiJ9...
        userId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: alice@example.com
        name:
          type: string
          example: Alice Smith
        registrationStage:
          $ref: '#/components/schemas/RegistrationStage'
        emailVerified:
          type: boolean
          example: false
        authProvider:
          type: string
          enum: [EMAIL, SPOTIFY]
          example: EMAIL
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
          nullable: true
          description: Required for EMAIL auth users; Spotify-only users may omit
          example: Secret@123

    # ── Onboarding Requests ────────────────────────────────────────────────
...
    PrivacySettingsRequest:
      type: object
      required: [isProfilePublic, showAge, showDistance, showLastActive, discoverable, showSpotifyProfile, showMusicStats, readReceipts]
      properties:
        isProfilePublic:
          type: boolean
          example: true
        showAge:
          type: boolean
          example: true
        showDistance:
          type: boolean
          example: true
        showLastActive:
          type: boolean
          example: false
        discoverable:
          type: boolean
          example: true
        showLikedByYou:
          type: boolean
          nullable: true
          example: false
        showSpotifyProfile:
          type: boolean
          example: true
        showMusicStats:
          type: boolean
          example: true
        incognitoMode:
          type: boolean
          nullable: true
          example: false
        readReceipts:
          type: boolean
          example: true

    # ── Onboarding Responses ───────────────────────────────────────────────
    UserDtoResponse:
      type: object
      description: Lean user response returned by all onboarding PUT endpoints
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Alice Smith
        email:
          type: string
          format: email
          example: alice@example.com
        imageUrl:
          type: string
          format: uri
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
          example: 45.8150
        longitude:
          type: number
          format: double
          nullable: true
          example: 15.9819
        city:
          type: string
          nullable: true
          example: Zagreb
        country:
          type: string
          nullable: true
          example: Croatia
        registrationStage:
          $ref: '#/components/schemas/RegistrationStage'
        gender:
          $ref: '#/components/schemas/Gender'
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: '1997-04-15'
        sexualOrientation:
          $ref: '#/components/schemas/SexualOrientation'
...
    PrivacySettingsResponse:
      type: object
      properties:
        isProfilePublic:
          type: boolean
        showAge:
          type: boolean
        showDistance:
          type: boolean
        showLastActive:
          type: boolean
        discoverable:
          type: boolean
        showLikedByYou:
          type: boolean
          nullable: true
        showSpotifyProfile:
          type: boolean
        showMusicStats:
          type: boolean
        incognitoMode:
          type: boolean
          nullable: true
        readReceipts:
          type: boolean
...
    UserProfileResponse:
      type: object
      description: |
        Public profile. Privacy-gated fields are null for non-owners when the corresponding
        privacy setting is disabled.
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
          example: Bob Jones
        age:
          type: integer
          nullable: true
          description: Null for non-owners when showAge=false
          example: 27
        gender:
          $ref: '#/components/schemas/Gender'
        locationCity:
          type: string
          nullable: true
          example: Zagreb
        locationCountry:
          type: string
          nullable: true
          example: Croatia
        photos:
          type: array
          description: Sorted by displayOrder
          items:
            type: string
            format: uri
        bio:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        mbti:
          $ref: '#/components/schemas/MBTI'
        lookingForText:
          type: string
          nullable: true
        favoriteQuote:
          type: string
          nullable: true
        conversationStarters:
          type: string
          nullable: true
        education:
          $ref: '#/components/schemas/EducationLevel'
        occupation:
          type: string
          nullable: true
        smokingHabits:
          $ref: '#/components/schemas/SmokingHabits'
        drinkingHabits:
          $ref: '#/components/schemas/DrinkingHabits'
        exerciseFrequency:
          $ref: '#/components/schemas/ExerciseFrequency'
        wantsKids:
          $ref: '#/components/schemas/KidsPreference'
        religion:
          $ref: '#/components/schemas/Religion'

    # ── Spotify (Authenticated) ────────────────────────────────────────────
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...